reclassify S3File as S3Manager::Carrierwave to better describe its purpose
