+++ app/models/assignment_file.rb
@@ -1,5 +1,5 @@
 class AssignmentFile < ActiveRecord::Base
-  include S3File
+  include S3Manager::Carrierwave
 
   attr_accessible :file, :filename, :filepath, :assignment_id
