+++ app/models/challenge_file.rb
@@ -1,5 +1,5 @@
 class ChallengeFile < ActiveRecord::Base
-  include S3File
+  include S3Manager::Carrierwave
 
   attr_accessible :file, :filename, :filepath, :challenge_id
