+++ app/models/grade_file.rb
@@ -1,5 +1,5 @@
 class GradeFile < ActiveRecord::Base
-  include S3File
+  include S3Manager::Carrierwave
 
   attr_accessible :file, :filename, :filepath, :grade_id
