+++ lib/s3_manager.rb
@@ -4,6 +4,7 @@ require_relative "s3_manager/encryption"
 require_relative "s3_manager/kms"
 require_relative "s3_manager/object_summary"
 require_relative "s3_manager/manager"
+require_relative "s3_manager/carrierwave"
 
 module S3Manager
 end
