+++ spec/active_record_spec_helper.rb
@@ -17,7 +17,7 @@ require "yaml"
 require "./lib/display_helpers"
 require "./lib/model_addons/advanced_rescue"
 require "./lib/model_addons/improved_logging"
-require "./lib/s3_file"
+require "./lib/s3_manager"
 require_relative "support/sorcery_stubbing"
 require_relative "support/file_helpers"
 require_relative "support/world"
